Finish refactoring preprocessing.
